+++ CHANGELOG.md
@@ -7,8 +7,12 @@ Add changes to the "Unreleased Changes" section. Once you create a version (and 
 to a section with the version name.
 
 ## Unreleased Changes
+
 * Add a new tab in ZATCA Phase 2 Business Settings for configuration of the tax account per company.
 * Create Tax Category, Sales Taxes and Charges template and Item tax template and link them to the tax account created on creating new ZATCA Business Settings
+* Set ZATCA Business Settings Fields to Be readonly only after onboarding except for system manager.
+  * Updated fields: Company, Unit Name, Unit Serial, Address, seller name, VAT Registration Number and additional ids.
+
 ## 0.26.0
 
 * Remove `Fatoora Server Url` field and replace it with `Fatoora Server` Select Field with 3 options `['Sandbox', 'Simulation', 'Production']`